feat: dispatch-parity quality improvements (#24)
* chore: remove slop from embedded skill assets

Replace "seamlessly" and "Leverage" with concrete language
in shipped skill markdown files.


* feat: dispatch-parity quality improvements

- Pin all GitHub Actions to full commit SHAs
- Add CODEOWNERS file
- Add Dependabot for go modules and github-actions
- Add concurrency control to CI/PR workflows
- Add CodeQL security scanning workflow
- Add govulncheck vulnerability scanning workflow
- Standardize golangci-lint config with 30+ linters
- Add dispatch-level linters (errname, exhaustive, forcetypeassert, etc.)
- Add gofumpt strict formatting checks
- Add deadcode detection
- Add cosign code signing to release workflow
- Add SBOM generation (SPDX + CycloneDX) to release workflow
- Add comprehensive README badges (CI, CodeQL, Go Report Card, etc.)


---------

+++ cli/src/internal/copilot/console_windows.go
@@ -31,7 +31,7 @@ type consoleHandles struct {
 
 // attachConsole opens CONIN$/CONOUT$ and uses SetStdHandle to make them
 // the process's standard handles. Child processes spawned via exec.Cmd
-// will then inherit real console handles that Node.js recognises as a TTY
+// will then inherit real console handles that Node.js recognizes as a TTY
 // (isTTY = true, columns/rows populated).
 //
 // Simply passing the CONOUT$ *os.File as cmd.Stdout is NOT enough —
@@ -57,9 +57,9 @@ func attachConsole() (*consoleHandles, error) {
 	h.origErr, _, _ = procGetStdHandle.Call(stdErrorHandle)
 
 	// Redirect to console
-	_, _, _ = procSetStdHandle.Call(stdInputHandle, uintptr(conin.Fd()))
-	_, _, _ = procSetStdHandle.Call(stdOutputHandle, uintptr(conout.Fd()))
-	_, _, _ = procSetStdHandle.Call(stdErrorHandle, uintptr(conout.Fd()))
+	_, _, _ = procSetStdHandle.Call(stdInputHandle, conin.Fd())
+	_, _, _ = procSetStdHandle.Call(stdOutputHandle, conout.Fd())
+	_, _, _ = procSetStdHandle.Call(stdErrorHandle, conout.Fd())
 
 	return h, nil
 }
